refactor(back): split monolithic files into single-responsibility modules
