+++ back/src/config/constants.ts
@@ -1,0 +1,8 @@
+export const MAX_MESSAGES_LOADED = 300;
+
+export const PING_INTERVAL_MS = 30000;
+export const RATE_LIMIT_FREE_MESSAGES = 5;
+export const COOLDOWN_INITIAL_SECONDS = 1;
+export const COOLDOWN_MULTIPLIER = 2;
+export const MAX_IMAGE_BYTES = 110 * 1024;
+export const MAX_TRIES_PER_GAME = 6;
