+++ back/src/config/env.ts
@@ -1,0 +1,16 @@
+import dotenv from 'dotenv';
+
+dotenv.config();
+
+export const env = {
+  port: Number(process.env.PORT) || 27025,
+  ntfyUrl: process.env.NTFY_URL,
+  db: {
+    host: process.env.DB_HOST,
+    user: process.env.DB_USER,
+    port: process.env.DB_PORT,
+    password: process.env.DB_PASSWORD,
+    database: process.env.DB_DATABASE,
+    charset: process.env.DB_CHARSET,
+  },
+} as const;
