+++ back/src/models/FullUser.ts
@@ -1,5 +1,6 @@
 import WS from 'ws';
 import { Server } from '@surtom/interfaces';
+import { COOLDOWN_INITIAL_SECONDS, COOLDOWN_MULTIPLIER } from '../config/constants.js';
 
 class FullUser {
   id: string;
@@ -18,8 +19,8 @@ class FullUser {
     this.connection = connection;
     this.messageCount = 0;
     this.lastMessageTimestamp = null;
-    this.messageCooldown = 1;
-    this.cooldownMultiplier = 2;
+    this.messageCooldown = COOLDOWN_INITIAL_SECONDS;
+    this.cooldownMultiplier = COOLDOWN_MULTIPLIER;
     this.listeningTypes = [];
     this.ip = ip;
   }
