+++ back/src/models/Player.ts
@@ -1,0 +1,9 @@
+export interface Player {
+  id: number;
+  username: string;
+  password: string;
+  sessionHash?: string;
+  registrationDate: Date;
+  isAdmin: number;
+  isBanned: number;
+}
